Simplify the circular reference solution by separating the execution out of the manager
