+++ bundle/src/main/java/com/adobe/acs/commons/oak/impl/EnsureOakIndex.java
@@ -18,16 +18,18 @@
 package com.adobe.acs.commons.oak.impl;
 
 import com.adobe.acs.commons.analysis.jcrchecksum.ChecksumGenerator;
+import com.adobe.acs.commons.oak.EnsureOakIndexManager;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.resource.ResourceResolverFactory;
 import org.apache.sling.commons.scheduler.ScheduleOptions;
 import org.apache.sling.commons.scheduler.Scheduler;
-import org.osgi.service.cm.Configuration;
-import org.osgi.service.cm.ConfigurationAdmin;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
 import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
+import org.osgi.service.component.annotations.ReferencePolicy;
+import org.osgi.service.component.annotations.ReferencePolicyOption;
 import org.osgi.service.metatype.annotations.AttributeDefinition;
 import org.osgi.service.metatype.annotations.Designate;
 import org.osgi.service.metatype.annotations.ObjectClassDefinition;
@@ -35,9 +37,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.jcr.RepositoryException;
-import java.io.IOException;
 import java.util.Collections;
-import java.util.Dictionary;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
@@ -106,7 +106,7 @@ public class EnsureOakIndex implements AppliableEnsureOakIndex {
     private Scheduler scheduler;
 
     @Reference
-    private ConfigurationAdmin configurationAdmin;
+    private EnsureOakIndexManager indexManager;
 
     private String ensureDefinitionsPath;
     private String oakIndexesPath;
@@ -133,16 +133,8 @@ public class EnsureOakIndex implements AppliableEnsureOakIndex {
         this.immediate = config.immediate();
 
         String[] ignoredProps = config.properties_ignore();
-        String[] indexManagerConfiguredIgnoreProperties = getIndexManagerConfiguredIgnoreProperties();
+        String[] indexManagerIgnoredProps = getIndexManagerConfiguredIgnoreProperties();
 
-        setIgnoredProperties(ignoredProps, indexManagerConfiguredIgnoreProperties);
-
-        if (this.immediate) {
-            apply(false);
-        }
-    }
-
-    private void setIgnoredProperties(String[] ignoredProps, String[] indexManagerIgnoredProps) {
         if (ignoredProps.length == 0) {
             // Legacy: check if EnsureOakIndexManagerImpl has this property configured -- https://github.com/Adobe-Consulting-Services/acs-aem-commons/issues/1966
             if (indexManagerIgnoredProps.length != 0) {
@@ -159,33 +151,18 @@ public class EnsureOakIndex implements AppliableEnsureOakIndex {
                         + "; please delete it");
             }
         }
+        if (this.immediate) {
+            apply(false);
+        }
     }
 
-    /**
-     * @return the ignore properties, using the ConfigurationAdmin to avoid having a circular reference between EnsureOakIndexManagerImpl and EnsureOakIndex OSGi services
-     */
+
     private String[] getIndexManagerConfiguredIgnoreProperties() {
-        try {
-             Configuration configuration = configurationAdmin.getConfiguration(EnsureOakIndexManagerImpl.class.getName());
-            if (configuration != null && configuration.getProperties() != null) {
-                return getIndexManagerConfiguredIgnoreProperties(configuration.getProperties());
-            }
-        } catch (IOException e) {
-            log.warn("Could not get ignored properties from index manager configuration", e);
+        if (indexManager instanceof EnsureOakIndexManagerImpl) {
+            EnsureOakIndexManagerImpl impl = (EnsureOakIndexManagerImpl) indexManager;
+            return impl.getIgnoredProperties();
         }
-        return new String[]{};
-    }
-
-    private static String[] getIndexManagerConfiguredIgnoreProperties(Dictionary<String, Object> properties) {
-        Object indexManagerIgnoredProps = properties.get(PROP_ADDITIONAL_IGNORE_PROPERTIES);
-        if (indexManagerIgnoredProps != null) {
-            if (indexManagerIgnoredProps instanceof String[]) {
-                return (String[]) indexManagerIgnoredProps;
-            } else if (indexManagerIgnoredProps instanceof String) {
-                return new String[]{(String) indexManagerIgnoredProps};
-            }
-        }
-        return new String[]{};
+        return new String[] {};
     }
 
     /**
