+++ bundle/src/main/java/com/adobe/acs/commons/oak/impl/EnsureOakIndexManagerImpl.java
@@ -71,9 +71,10 @@ import java.util.concurrent.CopyOnWriteArrayList;
 @Designate(ocd = EnsureOakIndexManagerImpl.Config.class)
 //@formatter:on
 public class EnsureOakIndexManagerImpl extends AnnotatedStandardMBean implements EnsureOakIndexManager, EnsureOakIndexManagerMBean {
-    private static final Logger log = LoggerFactory.getLogger(EnsureOakIndexManagerImpl.class);
 
     //@formatter:off
+    private static final String[] DEFAULT_ADDITIONAL_IGNORE_PROPERTIES = new String[]{};
+    private String[] additionalIgnoreProperties = DEFAULT_ADDITIONAL_IGNORE_PROPERTIES;
     public static final String PROP_ADDITIONAL_IGNORE_PROPERTIES = "properties.ignore";
 
     /**
@@ -94,22 +95,17 @@ public class EnsureOakIndexManagerImpl extends AnnotatedStandardMBean implements
         )
         String[] properties_ignore() default {};
 
+        String webconsole_configurationFactory_nameHint() default "Additional Ignore properties: {properties.ignore}";
+
     }
+    //@formatter:on
 
     // Disable this feature on AEM as a Cloud Service
     @Reference(target="(distribution=classic)")
     RequireAem requireAem;
 
-    @Reference(
-            cardinality = ReferenceCardinality.MULTIPLE,
-            policyOption = ReferencePolicyOption.GREEDY,
-            policy = ReferencePolicy.DYNAMIC,
-            fieldOption = FieldOption.UPDATE
-    )
-    // Thread-safe ArrayList to track EnsureIndex service registrations
-    private CopyOnWriteArrayList<AppliableEnsureOakIndex> ensureIndexes =
-            new CopyOnWriteArrayList<AppliableEnsureOakIndex>();
-    //@formatter:on
+    @Reference
+    EnsureOakIndexExecutor ensureOakIndexExecutor;
 
     public EnsureOakIndexManagerImpl() throws NotCompliantMBeanException {
         super(EnsureOakIndexManagerMBean.class);
@@ -120,20 +116,7 @@ public class EnsureOakIndexManagerImpl extends AnnotatedStandardMBean implements
      */
     @Override
     public final int ensureAll(boolean force) {
-        log.info("Applying all un-applied ensure index definitions");
-
-        int count = 0;
-        for (AppliableEnsureOakIndex index : this.ensureIndexes) {
-            if (!index.isApplied() || force) {
-                index.apply(force);
-                count++;
-                log.debug("Started applying index definition on [ {} ]", index);
-            } else {
-                log.debug("Skipping... [ {} ] is already applied.", index);
-            }
-        }
-
-        return count;
+        return ensureOakIndexExecutor.ensureAll(force);
     }
 
     /**
@@ -142,30 +125,7 @@ public class EnsureOakIndexManagerImpl extends AnnotatedStandardMBean implements
     @Override
     public final int ensure(final boolean force,
                             final String ensureDefinitionPath) {
-        int count = 0;
-        for (AppliableEnsureOakIndex index : this.ensureIndexes) {
-            if ((!index.isApplied() || force)
-                    && StringUtils.equals(ensureDefinitionPath, index.getEnsureDefinitionsPath())) {
-                index.apply(force);
-                count++;
-                log.debug("Started async job applying index definition for {}", index);
-            } else {
-                log.debug("Skipping... [ {} ] is already applied.", index);
-            }
-        }
-        return count;
-    }
-
-    protected final void bindAppliableEnsureOakIndex(AppliableEnsureOakIndex index) {
-        if (index != null && !this.ensureIndexes.contains(index)) {
-            this.ensureIndexes.add(index);
-        }
-    }
-
-    protected final void unbindAppliableEnsureOakIndex(AppliableEnsureOakIndex index) {
-        if (index != null && this.ensureIndexes.contains(index)) {
-            this.ensureIndexes.remove(index);
-        }
+       return ensureOakIndexExecutor.ensure(force, ensureDefinitionPath);
     }
 
     /**
@@ -177,33 +137,19 @@ public class EnsureOakIndexManagerImpl extends AnnotatedStandardMBean implements
     @Override
     @SuppressWarnings("squid:S1192")
     public final TabularData getEnsureOakIndexes() throws OpenDataException {
+        return ensureOakIndexExecutor.getEnsureOakIndexes();
+    }
 
-        final CompositeType configType = new CompositeType(
-                "Ensure Oak Index Configurations",
-                "Ensure Oak Index Configurations",
-                new String[]{"Ensure Definitions Path", "Oak Indexes Path", "Applied", "Immediate"},
-                new String[]{"Ensure Definitions Path", "Oak Indexes Path", "Applied", "Immediate"},
-                new OpenType[]{SimpleType.STRING, SimpleType.STRING, SimpleType.BOOLEAN, SimpleType.BOOLEAN});
-
-        final TabularDataSupport tabularData = new TabularDataSupport(new TabularType(
-                "Ensure Oak Index Configuration",
-                "Ensure Oak Index Configuration",
-                configType,
-                new String[]{"Ensure Definitions Path", "Oak Indexes Path"}));
+    @Activate
+    protected void activate(Config config) {
+        additionalIgnoreProperties = config.properties_ignore();
+    }
 
 
-        for (final AppliableEnsureOakIndex index : this.ensureIndexes) {
-            final Map<String, Object> data = new HashMap<String, Object>();
-
-            data.put("Ensure Definitions Path", index.getEnsureDefinitionsPath());
-            data.put("Oak Indexes Path", index.getOakIndexesPath());
-            data.put("Applied", index.isApplied());
-            data.put("Immediate", index.isImmediate());
-
-            tabularData.put(new CompositeDataSupport(configType, data));
-        }
-
-        return tabularData;
+    protected String[] getIgnoredProperties() {
+        return Optional.ofNullable(this.additionalIgnoreProperties)
+                .map(array -> Arrays.copyOf(array, array.length))
+                .orElse(DEFAULT_ADDITIONAL_IGNORE_PROPERTIES);
     }
 
 }
