+++ bundle/src/test/java/com/adobe/acs/commons/oak/impl/EnsureOakIndexTest.java
@@ -17,17 +17,14 @@
  */
 package com.adobe.acs.commons.oak.impl;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import javax.management.NotCompliantMBeanException;
 
-import org.apache.felix.utils.collections.MapToDictionary;
 import org.apache.sling.commons.scheduler.Scheduler;
 import org.apache.sling.testing.mock.sling.junit.SlingContext;
 import org.junit.Before;
@@ -40,8 +37,6 @@ import org.mockito.junit.MockitoRule;
 import com.adobe.acs.commons.analysis.jcrchecksum.ChecksumGenerator;
 import com.adobe.acs.commons.analysis.jcrchecksum.impl.ChecksumGeneratorImpl;
 import com.adobe.acs.commons.util.RequireAem;
-import org.osgi.service.cm.Configuration;
-import org.osgi.service.cm.ConfigurationAdmin;
 
 
 public class EnsureOakIndexTest {
@@ -68,27 +63,22 @@ public class EnsureOakIndexTest {
     @Before
     public void setup() {
         context.registerService(ChecksumGenerator.class, new ChecksumGeneratorImpl());
-
+        context.registerService(EnsureOakIndexExecutor.class, new EnsureOakIndexExecutor());
         context.registerService(Scheduler.class, scheduler);
         context.registerService(RequireAem.class, requireAem,"distribution","classic");
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_ENSURE_DEFINITIONS_PATH, "/apps/com/indexes");
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_OAK_INDEXES_PATH, "/oak:index");
     }
 
-    private void setupIndexManager(Object ignoreProperties) throws NotCompliantMBeanException, IOException {
+    private void setupIndexManager(String[] ignoreProperties) throws NotCompliantMBeanException {
         EnsureOakIndexManagerImpl indexManager = new EnsureOakIndexManagerImpl();
         Map<String,Object> props = new HashMap<>();
-        if (ignoreProperties != null) {
-            props.put(EnsureOakIndexManagerImpl.PROP_ADDITIONAL_IGNORE_PROPERTIES, ignoreProperties);
-        }
+        props.put(EnsureOakIndexManagerImpl.PROP_ADDITIONAL_IGNORE_PROPERTIES, ignoreProperties);
         context.registerInjectActivateService(indexManager, props );
-        ConfigurationAdmin configurationAdmin = context.getService(ConfigurationAdmin.class);
-        Configuration configuration = configurationAdmin.getConfiguration(indexManager.getClass().getName());
-        configuration.update(new MapToDictionary(props));
     }
 
     @Test
-    public void testWithLocalIgnoredProperties() throws NotCompliantMBeanException, IOException {
+    public void testWithLocalIgnoredProperties() throws NotCompliantMBeanException {
         setupIndexManager(new String[] {});
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_ADDITIONAL_IGNORE_PROPERTIES, new String[] {"localProp"});
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_IMMEDIATE, false);
@@ -99,7 +89,7 @@ public class EnsureOakIndexTest {
     }
 
     @Test
-    public void testLegacyMode() throws NotCompliantMBeanException, IOException {
+    public void testLegacyMode() throws NotCompliantMBeanException {
         setupIndexManager(new String[] {"globalSetting"});
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_IMMEDIATE, false);
         context.registerInjectActivateService(eoi, ensureOakIndexProperties);
@@ -109,7 +99,7 @@ public class EnsureOakIndexTest {
     }
 
     @Test
-    public void testMixedMode() throws NotCompliantMBeanException, IOException {
+    public void testMixedMode() throws NotCompliantMBeanException {
         setupIndexManager(new String[] {"globalSetting"});
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_ADDITIONAL_IGNORE_PROPERTIES, new String[] {"localProp"});
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_IMMEDIATE, false);
@@ -120,23 +110,7 @@ public class EnsureOakIndexTest {
     }
 
     @Test
-    public void testIndexManagerConfigurationSingleString() throws NotCompliantMBeanException, IOException {
-        setupIndexManager("globalSetting");
-        ensureOakIndexProperties.put(EnsureOakIndex.PROP_IMMEDIATE, false);
-        context.registerInjectActivateService(eoi, ensureOakIndexProperties);
-        assertTrue(eoi.getIgnoreProperties().contains("globalSetting"));
-    }
-
-    @Test
-    public void testIndexManagerConfigurationAsMultiString() throws NotCompliantMBeanException, IOException {
-        setupIndexManager(new String[] {"globalSetting"});
-        ensureOakIndexProperties.put(EnsureOakIndex.PROP_IMMEDIATE, false);
-        context.registerInjectActivateService(eoi, ensureOakIndexProperties);
-        assertTrue(eoi.getIgnoreProperties().contains("globalSetting"));
-    }
-
-    @Test
-    public void testIndexManagerConfigurationNotSet() throws NotCompliantMBeanException, IOException {
+    public void testIndexManagerConfigurationNotSet() throws NotCompliantMBeanException {
         setupIndexManager(null);
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_IMMEDIATE, false);
         ensureOakIndexProperties.put(EnsureOakIndex.PROP_ADDITIONAL_IGNORE_PROPERTIES, new String[] {"localProp"});
